# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://oask.ca/api"
        }
    ],
    "paths": {
...
        "/admin/ministries/limit/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Renvoie les minitéres suivant une limite",
                "operationId": "5bd1dcb999ca5ba7d8dab4fb92cf04e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/ministries/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Ajouter un ministère",
                "operationId": "91dbc8dd97a57127180445adb20f42ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "ministry_name": {
                                        "type": "string",
                                        "example": "Evangelisation"
                                    },
                                    "leader_name": {
                                        "type": "string",
                                        "example": "Andy Kasanda"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Un ministére puissant dans l'église où plusieurs âmes sont gagnées pour le christ"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/ministries/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Renvoie les informations d'un ministère par son ID",
                "operationId": "5e8eb8edb4d44fb5f20ff68e90b36a3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/ministries/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Editer les informations d'un ministère",
                "operationId": "7d50ba7418bb7b5fd07e4a77665dfe60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Evangelisation"
                                    },
                                    "leader_name": {
                                        "type": "string",
                                        "example": "Andy Kasanda"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Un ministére puissant dans l'église où plusieurs âmes sont gagnées pour le christ"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/ministries/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Active ou désactive l'état d'un ministère par son ID",
                "operationId": "0c546ebf95c320747fdac27409b45597",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/ministries/delete": {
            "post": {
                "tags": [
                    "ADMINISTRATION__MINISTERE"
                ],
                "summary": "Supprimer un ministere",
                "operationId": "cf4da4ae3a7561b3a84ce4ed5aa9b881",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}